design changes-13/11
Footer-I build a footer for example-with logo and links(shortcuts)

AddRecipePage-I changed the basic form - I added almost all the relevant functions - tags, categories, adding ingredients, adding substitutes and instructions.

In general, I combined the css and the js on the page of adding the recipe so that we stay with react only.

Also - changes in colors and shapes on the pages