17/11-design changes
homePage-I redesigned the categories in terms of greed, currently 4*4, gave relevant names and added a relevant image for each category.

loginPage-I changed the form so that there would be consistency between all the forms on the site, I added validation checks to the input and changed the image.

registrationPage-I changed the form so that there would be consistency between all the forms on the site, I added validation checks to the input and changed the image.