20/11-Design Changes and Building Pages
HomePage-I have changed the pictures and logos in the circles to more modern ones, black and white.

I have changed the visibility of the categories of food, to look better and more modern.

I have built a page for each category, so the user can nevigate between the home page and the category that he want to see.

In general-I made a few changes to the colors of headlines and etc.